compiler/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vlm_inspector_compiler"
version = "0.1.0"
description = "Compiler service: paragraph -> validated DSL."
requires-python = ">=3.12"
dependencies = [
    "vlm_inspector_shared",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",  # for sync migration runner
    "pyyaml>=6.0",
    "click>=8.1",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-postgresql>=6.0",
    "ruff>=0.4",
    "httpx>=0.27",
]

[project.scripts]
vlm-compile = "compiler.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["compiler"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
